requirements.txt
pycryptodome
setuptools~=57.0.0